Add option to replace zero character with capital O
The main and binary displays use monospace font, this tends to have a dot
in the middle of the zero character to distinguish between zero and
capital O. In this case there is no ambiguity, you know it is a zero.
It might look nicer without the dot, so add an option to replace zero
characters with capital O at the point the text is displayed.
The underlying data in disp_stack is not changed so there are no
knock-on effects.
Up the version to 2.5