Re-work out of range integer conversions
When switching from float to integer mode, the existing behaviour if the
value is outside the range of the current integer width/signedness is not
very helpful (limit to min/max range of the current settings). Also, the
behaviour is inconsistent with pasting in a value in integer mode.
Change the behaviour.

The new behaviour is to ignore the unsigned/signed setting, determine if
the value fits based on the integer width. A value in the range from
INT8_MIN up to UINT8_MAX requires width 8, a value in the range from
INT16_MIN up to UINT16_MAX requires width 16 etc. The width will be
automatically increased if needs be (with a notification). The unsigned/signed
setting will not be changed eg. if current settings are width 8, unsigned,
decimal, a value of -1 will give 255. The user can then change to signed if
they wish. If the value is negative and less than INT64_MIN, or positive and
greater than UINT64_MAX, return 0 (with warning) and leave the width unchanged.
Apply the same procedure when pasting in a value in integer mode.
Tidy up the Instructions and a few other comments while here.
Up the version to 2.6