Tweak fontsize handling primarily for GTK3
I tried it on a newer GTK3 setup and it didn't work so well, so fix up the fontsize
handling for GTK3. While here, extend the arbitrary range of allowed font sizes.
Also, make the fontsize used for the pending binary operator label configurable
(the only reason for changing this specific font in the first place was to make
the minus symbol a bit bigger).
Version 2.9.